+++ client/pages/home/index.js
@@ -1,4 +1,3 @@
-import promise from 'es6-promise';
 import React from 'react';
 import {render} from 'react-dom';
 import {combineReducers} from 'redux-immutable';
@@ -13,9 +12,6 @@ import HomePage from './HomePage';
 import {context} from '../../utils/config';
 
 import '../../common/scss/main.scss';
-
-// Promise 兼容性处理
-promise.polyfill();
 
 const routes = (
   <Route path="/" component={App}>