+++ client/pages/page-1/index.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import {render} from 'react-dom';
-import promise from 'es6-promise';
 import Root from '../../Root';
 import routes from './routes';
 import reducers from './reducers';
 import {context} from '../../utils/config';
 import '../../common/scss/main.scss';
 
-// Promise 兼容性处理
-promise.polyfill();
 
 render(
   <Root routes={routes} reducers={reducers} basename={`${context}/page1`}/>,