+++ webpack.config.dev.babel.js
@@ -1,6 +1,9 @@
 const path = require('path');
 const fs = require('fs');
 const webpack = require('webpack');
+const precss = require('precss');
+const autoprefixer = require('autoprefixer');
+const flexbugs = require('postcss-flexbugs-fixes');
 
 const hotMiddlewareScript = 'webpack-hot-middleware/client?path=/__webpack_hmr&timeout=20000&reload=true';
 const appPath = path.resolve(__dirname, 'public');
@@ -10,7 +13,7 @@ const nodeModules = path.resolve(__dirname, 'node_modules');
 // 如果没有二级路径区分，可以设为 '', 如 http://ft.jd.com
 const context = process.env.URL_CONTEXT;
 
-//判断 dll 文件是否已生成
+// 判断 dll 文件是否已生成
 let dllExist = false;
 try {
   fs.statSync(path.resolve(appPath, 'dll', 'vendor.dll.js'));
@@ -20,12 +23,13 @@ try {
 }
 
 const webpackConfig = {
+  cache: true, // 开启缓存,增量编译
   devtool: 'eval-source-map', // 生成 source map文件
   target: 'web', // webpack 能够为多种环境构建编译, 默认是 'web'，可省略 https://doc.webpack-china.org/configuration/target/
   resolve: {
-    //自动扩展文件后缀名
-    extensions: ['.js', '.css', '.scss', '.less', '.png', '.jpg', '.gif'],
-    //模块别名定义，方便直接引用别名
+    // 自动扩展文件后缀名
+    extensions: ['.js', '.scss', '.css', '.png', '.jpg', '.gif'],
+    // 模块别名定义，方便直接引用别名
     alias: {
       'react-router-redux': path.resolve(nodeModules, 'react-router-redux-fixed/lib/index.js'),
     },
@@ -38,10 +42,10 @@ const webpackConfig = {
 
   // 入口文件 让webpack用哪个文件作为项目的入口
   entry: {
-    home: ['./client/pages/home/index.js', hotMiddlewareScript],
-    about: ['./client/pages/about/index.js', hotMiddlewareScript],
-    page1: ['./client/pages/page-1/index.js', hotMiddlewareScript],
-    page2: ['./client/pages/page-2/index.js', hotMiddlewareScript],
+    home: ['babel-polyfill', './client/pages/home/index.js', hotMiddlewareScript],
+    about: ['babel-polyfill', './client/pages/about/index.js', hotMiddlewareScript],
+    page1: ['babel-polyfill', './client/pages/page-1/index.js', hotMiddlewareScript],
+    page2: ['babel-polyfill', './client/pages/page-2/index.js', hotMiddlewareScript],
   },
 
   // 出口 让webpack把处理完成的文件放在哪里
@@ -49,7 +53,7 @@ const webpackConfig = {
     // 编译输出目录, 不能省略
     path: path.resolve(appPath, 'dist'), // 打包输出目录（必选项）
     filename: '[name].bundle.js', // 文件名称
-    //资源上下文路径，可以设置为 cdn 路径，比如 publicPath: 'http://cdn.example.com/assets/[hash]/'
+    // 资源上下文路径，可以设置为 cdn 路径，比如 publicPath: 'http://cdn.example.com/assets/[hash]/'
     publicPath: `${context}/dist/`,
   },
 
@@ -60,29 +64,41 @@ const webpackConfig = {
         enforce: 'pre',
         test: /\.js$/,
         exclude: /node_modules/,
-        use: 'eslint-loader'
+        use: {
+          loader: 'eslint-loader',
+          options: {
+            configFile: '.eslintrc.js',
+            emitError: true, // 验证失败，终止
+          }
+        }
       },
       {
         enforce: 'pre',
         test: /\.scss/,
         exclude: /node_modules/,
         use: {
-          loader: 'sasslint-loader',
+          loader: 'sasslint-loader-vendor',
           options: {
             configFile: '.sass-lint.yml',
-            emitError: true,
-            failOnWarning: true
+            emitError: false,
+            failOnWarning: false
           }
         }
       },
       {
         test: /\.js$/,
         exclude: /node_modules/,
-        use: 'babel-loader',
+        use: {
+          loader: 'babel-loader',
+          options: {
+            cacheDirectory: true
+          }
+        }
       },
       // https://github.com/webpack/url-loader
       {
-        test: /\.(png|jpg|gif)$/,
+        test: /\.(png|jpe?g|gif)/,
+        exclude: /node_modules/,
         use: {
           loader: 'url-loader',
           options: {
@@ -100,15 +116,25 @@ const webpackConfig = {
         use: ['style-loader', {
           loader: 'css-loader',
           options: {
-            sourceMap: true,
             modules: true,
-            localIdentName: '[name]__[local]__[hash:base64:5]'
+            //localIdentName: '[name]__[local]__[hash:base64:5]',
+            getLocalIdent: (context, localIdentName, localName, options) => {
+              return `${context.resourcePath.split('/').slice(-2, -1)}-${localName}`;
+            }
           }
         }, {
           loader: 'postcss-loader',
           options: {
-            pack: 'cleaner',
             sourceMap: true,
+            plugins: [
+              precss,
+              flexbugs,
+              autoprefixer({
+                //flexbox: 'no-2009',
+                browsers: ['Explorer >= 9', 'Edge >= 12', 'Chrome >= 45', 'Firefox >= 38',
+                  'Android >= 4.4', 'iOS >=8', 'Safari >= 9']
+              })
+            ]
           }
         }],
       },
@@ -117,15 +143,25 @@ const webpackConfig = {
         use: ['style-loader', {
           loader: 'css-loader',
           options: {
-            sourceMap: true,
             modules: true,
-            localIdentName: '[name]__[local]__[hash:base64:5]'
+            //localIdentName: '[name]__[local]__[hash:base64:5]',
+            getLocalIdent: (context, localIdentName, localName, options) => {
+              return `${context.resourcePath.split('/').slice(-2, -1)}-${localName}`;
+            }
           }
         }, {
           loader: 'postcss-loader',
           options: {
-            pack: 'cleaner',
             sourceMap: true,
+            plugins: [
+              precss,
+              flexbugs,
+              autoprefixer({
+                //flexbox: 'no-2009',
+                browsers: ['Explorer >= 9', 'Edge >= 12', 'Chrome >= 45', 'Firefox >= 38',
+                  'Android >= 4.4', 'iOS >=8', 'Safari >= 9']
+              })
+            ]
           }
         }, {
           // Webpack loader that resolves relative paths in url() statements
@@ -154,13 +190,6 @@ const webpackConfig = {
     }),
     new webpack.LoaderOptionsPlugin({
       debug: true,
-      options: {
-        // eslint 配置
-        eslint: {
-          emitError: true, // 验证失败，终止
-          configFile: '.eslintrc.js'
-        },
-      }
     })
   ]
 };